# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Droits d'Auteur sur Blockchain",
        "version": "1.0.0",
        "description": "API permettant la gestion des droits d'auteur et la distribution des royalties via la blockchain Polygon"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Serveur de développement"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "name": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    }
                }
            },
            "Work": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "contentHash": {
                        "type": "string",
                        "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "royaltyRecipients": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "pattern": "^0x[a-fA-F0-9]{40}$"
                        }
                    },
                    "shares": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 10000
                        }
                    },
                    "metadata": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string"
                            },
                            "category": {
                                "type": "string"
                            },
                            "tags": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": ["Authentification"],
                "summary": "Inscription d'un nouvel utilisateur",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    }
                                },
                                "required": ["email", "password", "name"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/works": {
            "post": {
                "tags": ["Œuvres"],
                "summary": "Enregistrer une nouvelle œuvre",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Work"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Œuvre enregistrée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "tokenId": {
                                                    "type": "string"
                                                },
                                                "transactionHash": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/works/{tokenId}": {
            "get": {
                "tags": ["Œuvres"],
                "summary": "Récupérer les détails d'une œuvre",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "tokenId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'œuvre récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Work"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/royalties/distribute/{tokenId}": {
            "post": {
                "tags": ["Royalties"],
                "summary": "Distribuer les royalties pour une œuvre",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "tokenId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "description": "Montant en MATIC à distribuer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Royalties distribuées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "distributions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "recipient": {
                                                                "type": "string"
                                                            },
                                                            "amount": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                },
                                                "transactionHash": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}